bookmaker/requirements.txt
black
fastapi
grpcio-tools
pre-commit
pytest
pytest-mock
uvicorn
